feat: introduce Dioxus-based chat GUI with real-time input analysis and LLM responses.

+++ ifl_core/gui/src/main.rs
@@ -125,7 +125,7 @@ fn App() -> Element {
                 &id,
                 InputEvent::KeyDelete {
                     kind: DeleteKind::Backspace,
-                    count: diff,
+                    count: diff as u32,
                     ts,
                 },
             ) {
